d {assertive.strings}	R Documentation

Create regex for repeated digits

Description

Creates a regex string for repeated digits.

Usage

d(lo, hi = NA_integer_, optional = FALSE)

Arguments

lo
Minimum number of digits to match.

hi
Optional maximum number of digits to match.

optional
If TRUE, the digits are optional.

Value

A character vector of regexes.

Note

If hi is omitted, the returned regex will only match the exact number of digits given by lo.

Examples

## Not run: 
d(1:5)
d(1:5, 6:8)
d(0:2, Inf)

## End(Not run)

[Package assertive.strings version 0.0-3 Index]