assert_all_are_single_characters {assertive.strings}	R Documentation

Is the input a single character?

Description

Checks to see if he unput is a single character.

Usage

assert_all_are_single_characters(x, na_ignore = FALSE,
  severity = getOption("assertive.severity", "stop"))

assert_any_are_single_characters(x, na_ignore = FALSE,
  severity = getOption("assertive.severity", "stop"))

is_single_character(x, .xname)

Arguments

x
A character vector.

na_ignore
A logical value. If FALSE, NA values cause an error; otherwise they do not. Like na.rm in many stats package functions, except that the position of the failing values does not change.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_single_character returns TRUE when the input is a single character (as determined by nchar; this excludes NA). The assert_* functions return nothing but throw an error if the corresponding is_* function returns FALSE.

Note

The behaviour of this function has changed so that NA inputs now return NA rather than FALSE.

See Also

nchar

Examples

x <- c("", "a", "aa", NA)
is_single_character(x)

[Package assertive.strings version 0.0-3 Index]